Inliner: Load return values on the stack before jumping to post-call.
Previously, the inliner would emit the following sequence to
replace xRETURN instructions:

  1. STORE the return value
  2. DROP the rest of the stack
  3. JUMP to post-call

And at post-call, it would start by LOADing the return value back
on the stack.

This commit moves (and potentially duplicates) the LOAD instruction
to before JUMPing to post-call. In the common case where the stack
is empty at the point of the xRETURN (save for the return value
itself), we can then completely omit storing and loading the value.

This change recovers the optimized bytecode that the LoadDestination
commit destroyed in `InlinerTest.oldInlineHigherOrderTest(). The
changes in that commit make it much more frequent to have xRETURN
nodes in the middle of a method body, and hence the new inliner
behavior makes more sense.

Note: there already existed a comment in Inliner.inlineCallsite()
that pretended that loading the return value was part of the
replacement sequence for xRETURN. It was basically lying. The new
behavior is aligned with that comment.