Handle `If` in `BCodeBodyBuilder.genCond()`.
This is more for completeness of `genCond()`. It can now push down
the jumps inside all the control structures for which it makes
sense to do so. It typically applies when the condition of an `if`
or `while` is an `if..else` expression.

It is however not always a win. If the result of the nested `if`s
must be converted to a boolean anyway, doing this will result in
duplication of the conversion to boolean.

This is a backport of dotty's commit
https://github.com/lampepfl/dotty/commit/32c3395233aff768a04685a328d6030b0dc0df07