Remove a useless transform in LambdaLift, which damages patmat trees.
There is no need for LambdaLift to transform the shape

  Return(Block(stats, expr))

into

  Block(stats, Return(expr))

yet for some reason, it has done that since it was first
implemented in 184a284ccc9c842adaaec728914f11a2cad33c18.

Unfortunately, doing so damages tree shapes produced by the
pattern matcher, so that they cannot be recognized by the match
optimization in the back-end.

We remove that case to improve the generated code.