Prevent Erasure from pushing an adapt onto matchEnd labels.
Doing so damages the shapes produced by the pattern matcher, so
that they cannot be recognized by the match optimization in the
back-end.

To prevent that, we intercept shapes of the form

  Block(_, _: LabelDef)

in Erasure.typed1, to recursively adapt with a wildcard prototype,
and only then let the enclosing `typed` adapt the result.

Concretely, instead of ending up with

  Block(_, TypeApply(LabelDef(...), asInstanceOf, Foo))

we end up with

  TypeApply(Block(_, LabelDef(...)), asInstanceOf, Foo)

which leaves the Block and LabelDef together.