Only keep static members when looking up java static members
In Java code, `A.B` can refer to a static or non-static member `B`.
The Scala compiler models a selection of a static member as a selection
from the companion object `A`.

If the selection `A.B` (select from object `A`) doesn't yield a result,
`handleMissing` creates a `SelectFromTypeTree` in a second try.

Another special handling for Java is that `A.B` may refer to a static
member `B` defined in a parent of `A`. This is done by
`inCompanionForJavaStatic` in `typedSelect`. This lookup should only
find static members (fixed by this PR).

Before this PR, `A.B` for a non-static member B succeeded through
`inCompanionForJavaStatic`, which means the qualifier `A` remained
a reference to the synthetic companion object `A`. This means that
the we obtained the a type `A.B` instead of `A#B`.

After this PR, the `inCompanionForJavaStatic` lookup fails and we go
to `handleMissing`, which changes the resulting tree into a
`SelectFromTypeTree` (select from class A).