stop SpecializeTypes for scala3Defined
Scala 3.2.0 adds the scala.deriving.Mirror.fromProductTyped
method to its stdlib. This method is forced by
SpecializeTypes which in turn causes a chain of forces, leading
to scala.Tuple being reached in an upper bound of
Mirror.ProductOf. This causes an error for any code reading an
enum/case-class/sealed-trait from Scala 3. So do not force any
scala 3 method in this phase, as Scala 3 does not support
specialisation.

Alternatives include further delaying when errors are reported
for problematic definitions, (i.e. do not throw TypeError),
but this appeared to have more strange knock-on effects.