Fix accessibility case for static java members
Given

```
public class A {
  protected static class AI { }
}
public class B extends A {
  public static class BI extends AI { }
}
```

The owner of `AI` is the module class `A$`, the owner of `BI` is the
module class `B$`. When checking if the protected `AI` can be accessed
in `B$`, we need to navigate from `B$` to `B`. The inheritance chain
is not reflected in the module classes.