construct URL in manifestAt directly
Prevoisly, calling `.getResource` sometimes returned an URL to a different jar, if there was one on classpath containing a manifest. This was because `getResource` delegates to the parent CL first, and only calls `findResource` if no manifest was found. As `findResource` is `protected` in `ClassLoader` and only made `public` by `java.net.URLClassLoader`, we resort to creating the URL ourselves instead.

fixes scala/bug#12677