new LinkedHashMap/LinkedHashSet implementation
The mutable HashMap/HashSet has been rewroten and the
performance is better. So rewrote the LinkedHashMap and
LinkedHashSet also to improve performance. The detailed
data can be seen in the PR.
Most codes are same with HashMap/HashSet but some are
different:
1. To keep binary compatibility, only api in old solution
are updated.The two class LinkedHashMap/LinkedHashSet
still don't have parameters. hashcode can't be realized since
it needs a new iterator which will break binary compatibility.
2. Add specific method to handle the order when adding/removing
the entry.
3. other minor changes.

Signed-off-by: Liang Yan <ckgppl_yan@sina.cn>