Lint eta-expansions with impure subexpressions
In an eta-expansion `foo.bar(baz) _`, `foo` and `baz` are evaluated
eagerly, so the entire expression is equivalent to

```
{
  val eta$0 = foo
  val eta$1 = baz
  xs... => eta$0.bar(eta$1)(xs...)
}
```

which is different from `xs... => foo.bar(baz)(xs...)`.
I belive not many users are aware of this and it's a pitfall that can
lead to bugs that are difficult to track down.

The lint warning is not issued if `foo` and `baz` are pure.