Fix crasher with -Xasync, string switch + boxed unit
Pretty un-idiomatic code to trigger this! It was
the expansion of two macros in the user's code base.

This commit updates the ANF transform to deal with BoxedUnit
specifically -- we don't need to create a var to assign
the result of such an expression, but we also can't just
leave a Unit typed expression (the transformed Match) around.

In the test case, that caused a later phase to crash
when performing switch-on-string translation.