Extend reporting methods with CodeActions (code fixes)
Add a new `actions: List[CodeAction]` parameter to reporting methods.
The purpose is to pass these along to Zinc and BSP so that the
compiler can suggest code edits to the editors in a structural manner.

To showcase this feature, CodeActions are added for
* Deprecation of procedure syntax
* Deprecation of empty-paren method auto-application
* Deprecation of val in for-comprehension
* Deprecation of paren-less lambda parameter

Existing subclasses of FilteringReporter should continue to work,
though the compiler is generally not binary compatible between minor
versions.