onceop annotation and linter
**Problem**
There are certain classes, such as `Iterator` where
the number of times you can call a method is restricted to once.
This kind of restriction is currently not expressed as type,
and it can catch users off guard.

**Solution**
This introduces a new annotation called `@onceop(...)`,
which can mark methods that can be called once per instance
per onceop group. For example, all `Iterator` methods
except for `hasNext` and `next` are marked with `@onceop("default")`
such that calling `take(2)` twice would now result to a warning.

Since I am tracking this within a scope, this is more as a
lightweight check, and by no means exhaustive.
It doesn't check for example passing an iterator to a function,
which inside could make a onceop call or not.