affine annotation and linter
**Problem**
There are certain classes, such as `Iterator` where the number of times
you can call a method is restricted to once.
This kind of restriction is currently not expressed as type, and it can catch users off guard.

**Solution**
This introduces a new annotation `@affine`, which can mark a type as an
affine type which can be used at-most once.
For example, `Iterator` trait is now marked as affine, with
`hasNext` and `next` opted out using `@affineOptOut` annotation.

Since I am tracking this within a scope, this is more of a lightweight
lint, and by no means exhaustive.
It doesn't check for example passing an iterator to a function, which inside it
could make a method call (or not?!).