Remove unnecessary wildcards in covariant positions
Use explicit types instead of wildcards in covariant positions of self
types. The aim is to avoid these types as they introduce extra type
members when unpickled by TASTy Query/MiMa and possibly in Dotty. This
change does not seem to be strictly necessary but is a simple way to
avoid potential issues in the pickles Scala 3 pickles for the standard
library.

These where identified by TASTy MiMa filters in Dotty:
* https://github.com/lampepfl/dotty/blob/588a0b1e2928967273f8b1e96cd97006fd64e6b8/project/TastyMiMaFilters.scala#L23-L24
* https://github.com/lampepfl/dotty/blob/588a0b1e2928967273f8b1e96cd97006fd64e6b8/project/TastyMiMaFilters.scala#L48