Fixes issue with blocking and BatchingExecutor
BatchingExecutor has a thread local with internal runnables for
Future/Promise. In blockOn method this threadLocal was set to Nil and
if there are tasks to execute, it passed to Batch to execute immediately.
Howerver, run method of Batch class requires that thread local must be
null. It produces IllegalArgumentException and all unexecuted tasks are
lost at this point (having Promise/Future onComplete callback never
called, meaning - it will never be completed).

Fixes scala/bug#9304