Disable caching when parsing TASTy files to prevent stale TASTy files
By default, URLClassLoader's getResourceAsStream function uses a
URLConnection that allows caching.

That function was being used to load TASTy files during compilation.

This can cause problems when multiple compilations happen in the same
instance of the JVM. The first Scala 2 compilation caches the TASTy file.
Then the Scala 3 dependency is changed and recompiled, which leads to the
TASTy file changing. Then a later compilation recompiles the Scala 2 code
that depends on the aforementioned TASTy file. When that compilation retrieves
the TASTy file it gets the cached/stale TASTy file rather than the up to
date TASTy file.

This situation is pretty easy to get into if you have a codebase split
across Scala 2 and 3, are using a worker strategy to compile, and have lots
of small build targets.