Consistently mark class fields for mixed-in vals / vars mutable
Fields generated in classes for mixed-in vals and vars are now
directly marked MUTABLE when they are created in the Fields info
transformer.

The fields need to be mutable (also vor vals) because initialization
is performed outside the class constructor (`T$_setter_$foo_=` method).

Previously the fields were marked MUTABLE in the Fields tree
transformation. This could lead to incorrectly final fields in bytecode
when the info transformer ran (again) after the tree transform. Details
in on scala bug 12881.