Allow `-java-output-version` flag, matching Scala 3
Following on from https://github.com/scala/scala/pull/9982 and
discussion in https://github.com/scala/scala/pull/9982/files#r1434172490,
this change enables `-java-output-version` as an alternative name for
the `-release` flag in Scala 2, as in Scala 3 `-java-output-version` is
the preferred flag name (see https://github.com/lampepfl/dotty/pull/14606 ),
with the alias of `-release` still permitted.

Correspondingly, `-Xunchecked-java-output-version` is also included as an
alternative for `-target`, though note use of this flag is discouraged
in favour of `-java-output-version`/`-release` for Java 9 and above.