Emit missing `POP` in backend when accessing static fields
Since 2.13 we emit fields in objects as static. The backend thus
has some support for static fields.

ASTs that access a static field are usually limited to object
fields and only occur from within the module itself (`GETSTATIC`
in the getter method, `PUTSTATIC` in `clinit`). In this case the
qualifier of the field access AST is always a pure expression
and therefore nothing is emitted (`genLoadQualUnlessElidable`).

Compiler plugins can however generate ASTs where static fields
are accessed with an impure qualifier. This triggers a bug in
the backend where `genLoadQualUnlessElidable` loads the qualifier
reference onto the stack without consuming it. This leads to
either a crash in the backend or at runtime.

The test case is a macro that imitates such a scenario. Before
the fix it crashes at runtime ("Bad type on operand stack").