Desugar switchable matches with guards in async methods
The async transformation supports `Match` trees that are left in place
for the backend to emit a switch in bytecode.

Guards are desugared into a jump to the default case

    case 22 if p => 5
    case _       => 6

becomes

    case 22 => if (p) 5 else default()
    case _  => default: { 6 }

This pattern is unexpected for the async transform and leads to
invalid bytecode if the guard `p` has an await call.

The fix is to desugar switchable matches with guards in patmat
if the enclosing method has the `AsyncAttachment`.