In case apply rewrite, only issue deprecation when using companion
When a case apply is rewritten to the constructor, only warn about
the class being deprecated if the companion object was used directly.

```
@deprecated case class C(x: Int)
C(1) // warn
val Alias = C
Alias(1) // no warn
```

The warning was introduced in PRs 4955 and 5646.

Scala 2.12 doesn't perform the case apply rewrite if the receiver is
an alias, but that changed in 2.13 in PR 10560, where the workaround
for fixing issue 5626 was narrowed down.