Don't insert box instructions around label jumps
For match trees that are not eliminated during patmat, to be emitted
as switches in bytecode, make sure erasure doesn't introduce boxing
instructions around label jumps.

A label jump is a jump in bytecode, it doesn't produce a value on the
stack that can be boxed. If the label type would need boxing, nothing
needs to be done: the boxing instructions are already introduced at the
jump target, where the label is defined.