Type provided arguments in presence of arity mismatch
When typing an application, the typer first checks whether the application supplies the correct number of arguments. If not, it emits a "not enough arguments" error and moves on without checking whether the arguments have the correct type. In the usual batch compilation mode this is fine because we've already found an error.

But in the context of an IDE like Metals this is a problem. If we don't type the arguments their trees may remain untyped and IDE functionality like autocomplete and type hints won't work. This change addresses the most common presentation of this issue, which arises frequently when making a call to a higher-order function with multiple arguments, e.g.

```
class Foo {
  def Bar: Double = 2.0
  def Baz: Double = 1.0
}

def multiArg(f: Foo => Double, s: String): Nothing = ???
multiArg(_./*CURSOR*/) // importantly we want to see Bar and Baz as completions here
```

We want to see completions for the members of Foo as we type the call, without having to ensure the correct arity first.

Fixes scala/bug#12909