improvement (?): Allow for clearing JrtClassPath caches
In long running sessions of presentation compiler it seem like these caches can easily get corrupted and we longer can use the compiler within the current JVM or classloader.

I couldn't find the exact problem, but this is the only difference with Scala 2.12 and Scala 3, where those caches are not present. We don'tget those issues there.

I added a workaround in Metals for this, but it uses reflection quite heavily.

Alternatively, maybe this caching is not needed? I see that it's created only in Global once, so maybe we could cache it in Global instead?